mgr/volumes/nfs: Move nfs code out of volumes plugin
This PR is to support integration of dashboard with volumes/nfs plugin.

Fixes: https://tracker.ceph.com/issues/50080
(cherry picked from commit af9336316f978a9f67faafad2eb3e608ed2616b1)
