+++ src/pybind/mgr/nfs/__init__.py
@@ -1,0 +1,1 @@
+from .module import Module
